Splitting gulp file tasks into individual files for better maintainability.
