+++ gulpfile.js/index.js
@@ -1,0 +1,2 @@
+var requireDir = require('require-dir');
+requireDir('./tasks', {recurse: true});
