+++ gulpfile.js/tasks/build.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+
+gulp.task('build', [
+  'clean',
+  'html',
+  'styles',
+  'scripts',
+  'test'
+]);
