+++ gulpfile.js/tasks/clean.js
@@ -1,0 +1,8 @@
+var gulp = require('gulp');
+var del = require('del');
+
+gulp.task('clean', function(cb) {
+  del([
+    'app/tmp'
+  ], cb);
+});
