+++ gulpfile.js/tasks/scripts.js
@@ -1,0 +1,7 @@
+var gulp = require('gulp');
+var bundler = require('../lib/bundler');
+
+gulp.task('scripts', function(cb) {
+  process.env.BROWSERIFYSWAP_ENV = 'dist';
+  bundler(cb);
+});
