+++ gulpfile.js/tasks/styles.js
@@ -1,0 +1,13 @@
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var browserSync = require('browser-sync');
+var reload = browserSync.reload;
+
+gulp.task('styles', function() {
+  return gulp.src('./src/main.less')
+    .pipe($.less())
+    .pipe($.autoprefixer())
+    .pipe($.rename('bundle.css'))
+    .pipe(gulp.dest('./dist'))
+    .pipe(reload({ stream: true }));
+});
