+++ gulpfile.js/tasks/test.js
@@ -1,0 +1,6 @@
+var gulp = require('gulp');
+
+gulp.task('test', [
+  'jshint',
+  'mocha'
+]);
