+++ package.json
@@ -57,6 +57,7 @@
     "pioneer": "^0.11.0",
     "proxyquire": "^1.3.0",
     "request": "^2.51.0",
+    "require-dir": "^0.2.0",
     "selenium-webdriver": "^2.44.0",
     "sinon": "^1.12.2",
     "sinon-chai": "^2.6.0",
